tbdev turndown: remove now-unused code in tensorboard/uploader (#6705)
Resolves #6603

This removes a sizable chunk of the code in the `tensorboard/uploader`
subdirectory, now that we've disabled the TensorBoard.dev uploader in
#6707. It also removes several Pip package deps that were only needed
for the removed code.

The remaining code (`logdir_loader.py`, `upload_tracker.py`, `util.py`,
their tests, and the protos) is preserved because it is still depended
on either by internal code or by [Cloud AI
Platform](https://github.com/googleapis/python-aiplatform/blob/cd31c1306a9a00a01fbc1dda56fe99ed567a4cfb/google/cloud/aiplatform/tensorboard/uploader.py#L62-L65).

Note that for `util.py`, I've removed everything except `RateLimiter`
since that was the only symbol where I could find outside dependencies.
To simplify the test code I also folded `FakeTime` from `test_util.py`
into `util_test.py`.

Tested with an internal sync to confirm tests will still pass (once a
few other changes land).
